requirements.txt
torch==2.0.1
pyyaml==6.0
numpy==1.23.5
scikit-learn==1.2.0
matplotlib==3.5.1
pandas==1.4.0